End crudRedux project

+++ crudredux/src/components/Producto.js
@@ -2,13 +2,26 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { eliminarProductoAction } from '../actions/productos.actions';
 import { useDispatch } from 'react-redux';
-import { Swal } from 'sweetalert2';
+import Swal from 'sweetalert2';
 
 const Producto = ({ producto }) => {
   const dispatch = useDispatch();
 
   const eliminarProducto = id => {
-    dispatch(eliminarProductoAction(id));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then(result => {
+      if (result.value) {
+        Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+      }
+      dispatch(eliminarProductoAction(id));
+    });
   };
 
   return (
